requirements.txt
absl-py==1.3.0
actionlib==1.14.0
angles==1.9.13
apturl==0.5.2
astunparse==1.6.3
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
bcrypt==3.1.7
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
cachetools==5.2.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.5.2
catkin-pkg-modules==0.5.2
cbor==1.0.0
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.3
empy==3.3.2
entrypoints==0.3
fasteners==0.14.1
fastimport==0.9.8
flatbuffers==22.10.26
future==0.18.2
gast==0.4.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg===1.13.1-unknown
greenlet==0.4.15
grpcio==1.50.0
h5py==3.7.0
httplib2==0.14.0
hyperlink==19.0.0
idna==3.4
image-geometry==1.16.2
importlib-metadata==5.0.0
incremental==16.10.1
interactive-markers==1.12.0
joblib==1.2.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keras==2.11.0
keras-flops==0.1.2
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser_geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.6
lockfile==0.12.2
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.1.2
message-filters==1.15.15
monotonic==1.5
mpi4py==3.0.3
msgpack==0.6.2
netifaces==0.10.4
nose==1.3.7
numpy==1.23.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cudnn-cu11==8.6.0.163
oauthlib==3.2.2
olefile==0.46
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==21.3
paramiko==2.6.0
pexpect==4.6.0
pillow==10.2.0
protobuf==3.19.6
psutil==5.5.1
py-ubjson==0.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
reportlab==3.5.34
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
resource_retriever==1.12.7
roman==2.0.0
rosbag==1.15.15
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.22.1
rosdep-modules==0.22.1
rosdistro==0.9.0
rosdistro-modules==0.9.0
rosgraph==1.15.15
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.15.15
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.15
rosmake==1.15.8
rosmaster==1.15.15
rosmsg==1.15.15
rosnode==1.15.15
rosparam==1.15.15
rospkg==1.4.0
rospkg-modules==1.4.0
rospy==1.15.15
rosservice==1.15.15
rostest==1.15.15
rostopic==1.15.15
rosunit==1.15.8
roswtf==1.15.15
rqt-moveit==0.5.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_image_view==0.4.16
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.3
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rviz==1.14.19
scikit-learn==1.3.2
scipy==1.9.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
ssh-import-id==5.10
systemd-python==234
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-gpu==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-ranking==0.5.1
tensorflow-serving-api==2.11.0
termcolor==2.1.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
threadpoolctl==3.1.0
topic-tools==1.15.15
Twisted==18.9.0
txaio==2.10.0
typing_extensions==4.4.0
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
vcstools==0.1.42
wadllib==1.3.3
Werkzeug==2.2.2
wrapt==1.14.1
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.14
xkit==0.0.0
zipp==3.10.0
zope.interface==4.7.1
